Moved non-Variable tests to test_misc.py
